:recycle: move buildVersion to shared
