+++ api/tests/certification/evaluation/unit/domain/services/scoring/calibrated-challenge-service_test.js
@@ -46,7 +46,7 @@ describe('Certification | Evaluation | Unit | Domain | Services | calibrated cha
         const certificationCourseId = 1234;
         const assessmentId = 5678;
 
-        const version = domainBuilder.certification.evaluation.buildVersion({
+        const version = domainBuilder.certification.shared.buildVersion({
           startDate: new Date('2025-06-22'),
         });
         calibratedChallengeRepository.getAllCalibratedChallenges.withArgs({ version }).resolves(challengeList);
@@ -93,7 +93,7 @@ describe('Certification | Evaluation | Unit | Domain | Services | calibrated cha
         // given
         const certificationCourseId = 1234;
         const assessmentId = 5678;
-        const version = domainBuilder.certification.evaluation.buildVersion({
+        const version = domainBuilder.certification.shared.buildVersion({
           startDate: new Date('2025-06-22'),
         });
         calibratedChallengeRepository.getAllCalibratedChallenges.withArgs({ version }).resolves(challengeList);
