+++ api/tests/certification/evaluation/unit/domain/services/scoring/scoring-v3_test.js
@@ -76,7 +76,7 @@ describe('Certification | Evaluation | Unit | Domain | Services | Scoring V3', f
         findByCertificationCourseAndVersion: sinon.stub(),
       };
 
-      version = domainBuilder.certification.evaluation.buildVersion({
+      version = domainBuilder.certification.shared.buildVersion({
         challengesConfiguration: {
           maximumAssessmentLength: 1,
           defaultCandidateCapacity: -3,
@@ -154,7 +154,7 @@ describe('Certification | Evaluation | Unit | Domain | Services | Scoring V3', f
         });
 
         const candidate = domainBuilder.buildCertificationCandidate({ reconciledAt: new Date('2021-01-01') });
-        const version = domainBuilder.certification.evaluation.buildVersion({
+        const version = domainBuilder.certification.shared.buildVersion({
           challengesConfiguration: {
             maximumAssessmentLength: 1,
             defaultCandidateCapacity: -3,
@@ -358,7 +358,7 @@ describe('Certification | Evaluation | Unit | Domain | Services | Scoring V3', f
             });
 
             const candidate = domainBuilder.buildCertificationCandidate({ reconciledAt: new Date('2021-01-01') });
-            const version = domainBuilder.certification.evaluation.buildVersion({
+            const version = domainBuilder.certification.shared.buildVersion({
               challengesConfiguration: {
                 maximumAssessmentLength: 1,
                 defaultCandidateCapacity: -3,
@@ -726,7 +726,7 @@ describe('Certification | Evaluation | Unit | Domain | Services | Scoring V3', f
           certificationCourseId,
         });
         candidate = domainBuilder.buildCertificationCandidate({ reconciledAt: new Date('2021-01-01') });
-        version = domainBuilder.certification.evaluation.buildVersion({
+        version = domainBuilder.certification.shared.buildVersion({
           challengesConfiguration: {
             maximumAssessmentLength: 1,
             defaultCandidateCapacity: -3,
