+++ api/tests/certification/evaluation/unit/domain/usecases/simulate-flash-assessment-scenario_test.js
@@ -13,7 +13,7 @@ describe('Unit | Domain | Usecases | simulate-flash-assessment-scenario', functi
       const accessibilityAdjustmentNeeded = false;
       const versionId = 1;
       const challengesConfiguration = { minimumEstimatedSuccessRateRanges: [], defaultCandidateCapacity: -1 };
-      const version = domainBuilder.certification.evaluation.buildVersion({
+      const version = domainBuilder.certification.shared.buildVersion({
         id: versionId,
         scope: Scopes.PIX_PLUS_DROIT,
         challengesConfiguration,
